fix for memcached-0.14
git-svn-id: http://svn.thrudb.org/thrudb/trunk@215 a07ddc84-1b2c-4fb2-82be-f96b6b69ae06

+++ thrudoc/src/MemcachedBackend.cpp
@@ -44,7 +44,7 @@ string MemcachedBackend::get (const string & bucket, const string & key )
     memcached_st * cache = get_cache ();
     string cache_key = (bucket + ":" + key);
     memcached_return rc;
-    uint16_t opt_flags = 0;
+    uint32_t opt_flags = 0;
     size_t str_length;
     char * str = memcached_get(cache, (char*)cache_key.c_str (),
                                cache_key.length (), &str_length,
